requirements.txt
streamlit
pandas
matplotlib
bcrypt
fpdf
python-dateutil